# fastapi-server/src/openapi_server/models/paginated_result.py
# coding: utf-8

"""
    Payload API

    An API to manage transportation payload with the the following operations: * List all Payloads * Create a Payload (aka Route) * Get Payload Estimate * Update Payload with Actual * Delete a Payload  Features: * Result pagination * Security (Auth'n / Auth'z)  We will use a simple **Payload** resource representation with the following properties: * ID (Route, aka Flight) * Passengers (count) * Baggage (weight) * Cargo (weight) * EstimateDateTimestamp (optional) * LastUpdateTimestamp (optional) 

    The version of the OpenAPI document: v1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictInt
from typing import Any, ClassVar, Dict, List, Optional
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class PaginatedResult(BaseModel):
    """
    PaginatedResult
    """ # noqa: E501
    offset: Optional[StrictInt] = Field(default=None, description="Page offset")
    limit: Optional[StrictInt] = Field(default=None, description="Page size limit")
    total: Optional[StrictInt] = Field(default=None, description="Total number of records across all pages")
    data: Optional[List[Any]] = None
    __properties: ClassVar[List[str]] = ["offset", "limit", "total", "data"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of PaginatedResult from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of PaginatedResult from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "offset": obj.get("offset"),
            "limit": obj.get("limit"),
            "total": obj.get("total"),
            "data": obj.get("data")
        })
        return _obj